+++ src/Timezones.spec.ts
@@ -7,28 +7,39 @@ import {
 
 describe("Timezones", () => {
   describe("extractOffsetMinutes", () => {
-    const arr = [{ tz: "7", exp: "UTC+7" }, { tz: "3:30", exp: "UTC+03:30" }]
-    ;[
+    function ozn(tz: string) {
+      return {
+        tz,
+        src: "offsetMinutesToZoneName"
+      }
+    }
+    const arr = [{ s: "7", exp: "UTC+7" }, { s: "3:30", exp: "UTC+03:30" }]
+    const ex = [
       { tz: "", exp: undefined },
       { tz: "garbage", exp: undefined },
-      { tz: "09:00", exp: "UTC+9" },
-      { tz: "America/Los_Angeles", exp: "America/Los_Angeles" },
-      ...arr,
-      ...arr.map(({ tz, exp }) => ({ tz: "+" + tz, exp })),
-      ...arr.map(({ tz, exp }) => ({ tz: "UTC+" + tz, exp })),
-      ...arr.map(({ tz, exp }) => ({
-        tz: "-" + tz,
-        exp: exp.replace("+", "-")
+      { tz: "09:00", exp: ozn("UTC+9") },
+      {
+        tz: "America/Los_Angeles",
+        exp: { tz: "America/Los_Angeles", src: "validIANAZone" }
+      },
+      ...arr.map(({ s, exp }) => ({ tz: s, exp: ozn(exp) })),
+      ...arr.map(({ s, exp }) => ({ tz: "+" + s, exp: ozn(exp) })),
+      ...arr.map(({ s, exp }) => ({ tz: "UTC+" + s, exp: ozn(exp) })),
+      ...arr.map(({ s, exp }) => ({
+        tz: "-" + s,
+        exp: ozn(exp.replace("+", "-"))
       })),
-      ...arr.map(({ tz, exp }) => ({
-        tz: "UTC-" + tz,
-        exp: exp.replace("+", "-")
+      ...arr.map(({ s, exp }) => ({
+        tz: "UTC-" + s,
+        exp: ozn(exp.replace("+", "-"))
       }))
-    ].forEach(({ tz, exp }) =>
+    ]
+
+    for (const { tz, exp } of ex) {
       it(`("${tz}") => ${exp}`, () => {
         expect(extractOffset(tz)).to.eql(exp)
       })
-    )
+    }
   })
 
   describe("extractTzOffsetFromTags", () => {
@@ -40,13 +51,22 @@ describe("Timezones", () => {
         { tzo: "+02:00", exp: "UTC+2" }
       ]) {
         it(`({ TimeZone: ${tzo}}) => ${exp}`, () => {
-          expect(extractTzOffsetFromTags({ TimeZone: tzo })).to.eql(exp)
+          expect(extractTzOffsetFromTags({ TimeZone: tzo })).to.eql({
+            tz: exp,
+            src: "offsetMinutesToZoneName from TimeZone"
+          })
         })
         it(`({ OffsetTime: ${tzo}}) => ${exp}`, () => {
-          expect(extractTzOffsetFromTags({ OffsetTime: tzo })).to.eql(exp)
+          expect(extractTzOffsetFromTags({ OffsetTime: tzo })).to.eql({
+            tz: exp,
+            src: "offsetMinutesToZoneName from OffsetTime"
+          })
         })
         it(`({ TimeZoneOffset: ${tzo}}) => ${exp}`, () => {
-          expect(extractTzOffsetFromTags({ TimeZoneOffset: tzo })).to.eql(exp)
+          expect(extractTzOffsetFromTags({ TimeZoneOffset: tzo })).to.eql({
+            tz: exp,
+            src: "offsetMinutesToZoneName from TimeZoneOffset"
+          })
         })
       }
     })
@@ -58,7 +78,10 @@ describe("Timezones", () => {
           CreateDate: "2014:07:19 12:05:19",
           DateTimeUTC: "2014:07:19 19:05:19"
         })
-      ).to.eql("UTC-7")
+      ).to.eql({
+        tz: "UTC-7",
+        src: "offset between CreateDate and DateTimeUTC"
+      })
     })
     it("with DateTimeUTC and very different created-at DateTime", () => {
       expect(